requirements.txt
blinker==1.9.0
click==8.3.0
colorama==0.4.6
dnspython==2.8.0
Flask==3.1.2
Flask-Login==0.6.3
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==25.0
psycopg2-binary==2.9.11
pymongo==4.15.5
python-dotenv==1.2.1
Werkzeug==3.1.3
Flask-WTF==1.2.2
Flask-Limiter==4.1.1
pysnmp>=7.0.0
pysmi>=1.6.0
